Move ethereum tests to their own folder
