+++ libs/ledger-live-common/src/families/ethereum/__tests__/hw-signMessage.test.ts
@@ -1,11 +1,11 @@
-import "../../__tests__/test-helpers/setup";
+import "../../../__tests__/test-helpers/setup";
 
+import { fail } from "assert";
 import { EIP712Message } from "@ledgerhq/hw-app-eth/lib/modules/EIP712";
 import testEIP712Message from "@ledgerhq/hw-app-eth/tests/fixtures/messages/0.json";
-import ethSign from "./hw-signMessage";
-import { createFixtureCryptoCurrency } from "../../mock/fixtures/cryptoCurrencies";
+import { createFixtureCryptoCurrency } from "../../../mock/fixtures/cryptoCurrencies";
 import { StatusCodes, TransportStatusError } from "@ledgerhq/errors";
-import { fail } from "assert";
+import ethSign from "../hw-signMessage";
 
 const signPersonalMessage = jest.fn(() =>
   Promise.resolve({
