+++ libs/ledger-live-common/src/families/ethereum/__tests__/synchronisation.integration.test.ts
@@ -1,13 +1,13 @@
-import "../../__tests__/test-helpers/setup";
+import "../../../__tests__/test-helpers/setup";
 import { reduce } from "rxjs/operators";
-import { fromAccountRaw } from "../../account";
-import { getAccountCurrency } from "../../account/helpers";
+import { fromAccountRaw } from "../../../account";
+import { getAccountCurrency } from "../../../account/helpers";
 import type { Account, SubAccount, OperationRaw } from "@ledgerhq/types-live";
-import { getAccountBridge } from "../../bridge";
-import { makeBridgeCacheSystem } from "../../bridge/cache";
-import { ethereum1 } from "./datasets/ethereum1";
-import { ethereum2 } from "./datasets/ethereum2";
-import { encodeOperationId } from "../../operation";
+import { getAccountBridge } from "../../../bridge";
+import { makeBridgeCacheSystem } from "../../../bridge/cache";
+import { ethereum1 } from "../datasets/ethereum1";
+import { ethereum2 } from "../datasets/ethereum2";
+import { encodeOperationId } from "../../../operation";
 import {
   getCryptoCurrencyById,
   listTokensForCryptoCurrency,
