+++ libs/ledger-live-common/src/platform/logic.test.ts
@@ -168,7 +168,7 @@ describe("completeExchangeLogic", () => {
         nonce: 8,
         data: Buffer.from("Some data...", "hex"),
         gasPrice: new BigNumber("700000"),
-        maxBaseFeePerGas: null,
+        maxFeePerGas: null,
         maxPriorityFeePerGas: null,
         userGasLimit: new BigNumber("1200000"),
         feesStrategy: "medium",
